fix: carousel hitbox edges fix (COR-4364) (#486)
**Fixes or implements VF-4364**

### Brief description. What is this change?

Fixed the carousel button hitbox issue by making the 'surrounding' button component a little bit bigger so when the image is enlarged it will still be inside of the 'hitbox'.

Added the `NewChat` component wrapper to the Carousel story, so that it looks good in storybook.
Added a `2px` left margin to the first carousel card to align it with the other messages displayed above or below it.

Also updated `@vanilla-extract/css` and `@vanilla-extract/vite` packages.
Initially did this cause I thought there was a vanilla-extract related bug - there wasn't, but the update is still helpful to stay updated.